requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.3
cuda-bindings==13.1.1
cuda-pathfinder==1.3.3
cuda-python==13.1.1
cycler==0.12.1
defusedxml==0.7.1
fastapi==0.128.0
filelock==3.20.3
flatbuffers==25.12.19
fonttools==4.61.1
fsspec==2026.1.0
ftfy==6.3.1
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.3.7
humanfriendly==10.0
idna==3.11
Jinja2==3.1.6
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.8
ml_dtypes==0.5.4
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.2
onnx==1.20.1
onnxruntime==1.23.2
onnxslim==0.1.82
open_clip_torch==3.2.0
opencv-python==4.13.0.90
opencv-python-headless==4.13.0.90
packaging==26.0
pillow==12.1.0
polars==1.37.1
polars-runtime-32==1.37.1
protobuf==6.33.5
psutil==7.2.2
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.22
PyYAML==6.0.3
regex==2026.1.15
requests==2.32.5
safetensors==0.7.0
scipy==1.17.0
setuptools==80.10.2
shellingham==1.5.4
six==1.17.0
starlette==0.50.0
supervision==0.27.0
sympy==1.14.0
tensorrt==10.13.3.9
tensorrt_dispatch==10.13.3.9
tensorrt_lean==10.13.3.9
timm==1.0.24
torch==2.10.0
torchvision==0.25.0
tqdm==4.67.2
typer-slim==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
ultralytics==8.4.10
ultralytics-thop==2.0.18
urllib3==2.6.3
uvicorn==0.40.0
wcwidth==0.5.3
